requirements.txt
tensorflow
bert-serving-server
bert-serving-client
sklearn
pytest
pytest-xdist
logzero